100% background fix, started c-plugin examples
Fix: When displaying a 100% value, the background circle stroke was not drawn due to the assumtion that the foreground pie would totally cover it. But this assuption was wrong in case the foreground graphic was a ring and the background used a strokeWidth > ringWidth.

Started new examples page for svgContentPlugins.