refactorings: -gapToRing, +margin
introduced new margin property for contentPlugins, changed getBackgroundRadius() API function and removed old gapToRing option. The new margin is now also applicable to pie charts and not restricted to rings. getBackgroundRadius() now does not depend on a background color being set. Plug-ins that want to ignore the margin in cases like no background-color have to declare this with a boolean argument to the function.