FIX: CheckComplete-Plug-In in CSS Mode…
The check icon was drawn too large in CSS mode, if the fullSize option was set and the cssClassBackgroundCircle option was used instead of the backgroundColor option.

(If drawn on a background circle, the icon is smaller than if drawn into the blank space of a ring chart. Due to the bug, the missing backgroundColor option was misinterpreted as "no background circle" during icon size calculations, since the new cssClassBackgroundCircle option was ignored in that case.)

Also added a test case for this scenario to examplesContentPlugins.html.